# Conflicts:
#	crm/src/main/java/com/hy/crm/controller/UserController.java
#	crm/src/main/java/com/hy/crm/entity/Business.java
#	crm/src/main/java/com/hy/crm/entity/Clien.java
#	crm/src/main/java/com/hy/crm/entity/Data.java
#	crm/src/main/java/com/hy/crm/mapper/DataMapper.java
#	crm/src/main/java/com/hy/crm/mapper/DocumentaryMapper.java
#	crm/src/main/java/com/hy/crm/service/impl/DataServiceImpl.java
#	crm/src/main/java/com/hy/crm/util/FileUpload.java
#	crm/src/main/java/com/hy/crm/util/LayuiData.java
许福琦 第一次整合

+++ crm/src/main/java/com/hy/crm/controller/After_saleController.java
@@ -10,17 +10,13 @@ import com.hy.crm.service.impl.ClienServiceImpl;
 import com.hy.crm.service.impl.ContractServiceImpl;
 import com.hy.crm.service.impl.DataServiceImpl;
 import com.hy.crm.util.LayuiData;
-import net.sf.jsqlparser.statement.select.Limit;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
-
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
-import java.util.zip.DataFormatException;
 
 /**
  * <p>
@@ -57,7 +53,7 @@ public class After_saleController {
      */
     @RequestMapping("/statistics.do")
     @ResponseBody
-    public List<Long> statistics(){
+    public List<Integer> statistics(){
         return after_saleService.counts();
     }
